requirements.txt
langchain
huggingface_hub
openai
uvicorn
fastapi
scikit-learn
psycopg2-binary